move implementation do DeviceSession
