+++ packages/vscode-extension/src/project/project.ts
@@ -492,18 +492,10 @@ export class Project implements Disposable, ProjectInterface, DeviceSessionsMana
   // #region File Transfer
 
   public async openSendFileDialog() {
-    const pickerResult = await window.showOpenDialog({
-      canSelectMany: true,
-      canSelectFolders: false,
-      title: "Select files to send to device",
-    });
-    if (!pickerResult) {
-      return; // no files selected
+    if (!this.deviceSession) {
+      throw new Error("No device session available");
     }
-    const sendFilePromises = pickerResult.map((fileUri) => {
-      return this.deviceSession?.sendFile(fileUri.fsPath);
-    });
-    await Promise.all(sendFilePromises);
+    this.deviceSession.openSendFileDialog();
   }
 
   public async sendFileToDevice({
@@ -516,17 +508,7 @@ export class Project implements Disposable, ProjectInterface, DeviceSessionsMana
     if (!this.deviceSession) {
       throw new Error("No device session available");
     }
-    const tempDir = await fs.promises.mkdtemp(os.tmpdir());
-    try {
-      const tempFileLocation = path.join(tempDir, fileName);
-      await fs.promises.writeFile(tempFileLocation, new Uint8Array(data));
-      await this.deviceSession.sendFile(tempFileLocation);
-    } finally {
-      // NOTE: no `await` here, this can safely go in the background
-      fs.promises.rm(tempDir, { recursive: true }).catch((_e) => {
-        /* silence the errors, it's fine */
-      });
-    }
+    this.deviceSession.sendFileToDevice(fileName, data);
   }
 
   // #endregion
